requirements.txt
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jsonapi==2.4.0
inflection==0.3.1
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
